requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.0.0=pypi_0
astunparse=1.6.3=pypi_0
bzip2=1.0.8=he774522_0
ca-certificates=2023.12.12=haa95532_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pypi_0
cffi=1.16.0=py311h2bbff1b_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=py311haa95532_0
colorama=0.4.6=py311haa95532_0
cryptography=41.0.7=py311h89fc84f_0
flask=2.2.5=py311haa95532_0
flask-cors=4.0.0=pypi_0
flatbuffers=23.5.26=pypi_0
freetype=2.12.1=ha860e81_0
gast=0.4.0=pypi_0
giflib=5.2.1=h8cc25b3_3
google-auth=2.25.2=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.60.0=pypi_0
h5py=3.10.0=pypi_0
idna=3.6=pypi_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.1.2=py311haa95532_0
joblib=1.3.2=pypi_0
jpeg=9e=h2bbff1b_1
keras=2.13.1=pypi_0
lerc=3.0=hd77b12b_0
libclang=16.0.6=pypi_0
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libwebp=1.3.2=hbc33d0d_0
libwebp-base=1.3.2=h2bbff1b_0
lz4-c=1.9.4=h2bbff1b_0
markdown=3.5.1=pypi_0
markupsafe=2.1.3=py311h2bbff1b_0
ml-dtypes=0.2.0=pypi_0
numpy=1.24.3=pypi_0
oauthlib=3.2.2=pypi_0
opencv-python=4.8.1.78=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.12=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=23.2=pypi_0
pillow=10.0.1=py311h045eedc_0
pip=23.3.2=pypi_0
protobuf=4.23.4=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pymysql=1.0.2=py311haa95532_1
pyodbc=4.0.39=py311hd77b12b_0
python=3.11.5=he1021f5_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-learn=1.3.1=pypi_0
scipy=1.11.4=pypi_0
setuptools=68.2.2=py311haa95532_0
six=1.16.0=pypi_0
sqlite=3.41.2=h2bbff1b_0
tensorboard=2.13.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
tensorflow=2.13.0=pypi_0
tensorflow-estimator=2.13.0=pypi_0
tensorflow-intel=2.13.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.4.0=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h2bbff1b_0
typing-extensions=4.5.0=pypi_0
tzdata=2023c=h04d1e81_0
urllib3=2.1.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py311haa95532_0
wrapt=1.14.1=pypi_0
xz=5.4.5=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
